requirements.txt
Django==4.2.1
django-countries==7.5.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Faker~=18.9.0
factory_boy==3.2.1
django_filter==23.2
martor
gunicorn
Pillow